+++ MoonBot/release_notes.md
@@ -1,3 +1,8 @@
+MoonBot V0.0.3
+==============
+* 1.add rover follow ball in `MoonBot_Rover.h`
+* 2.add new humannoid actions `node()`
+
 MoonBot V0.0.2
 ==============
 * 1.include `MoonBot_Humannoid.h` in file `MoonBot.h`