+++ release_notes.md
@@ -1,3 +1,11 @@
+MoonBot-Mixly V1.0.4
+====================
+* 1.update MoonBot Arduino library to V1.0.3
+* 2.add Blocks: MECH/Robot
+* 3.reverse IR/touch sensor signal
+* 4.bug fixes, now speaker voice `Cat` works correctly
+* 5.bug fixes, now `moonbot.nod()` corks correctly
+
 MoonBot-Mixly V0.0.2
 ====================
 * 1.Bug fixes, now set servo angle block works correctly.